In which I spent an evening plotting a bitmap font
The idea here is to eventually write something that can convert these
(very simple to edit) grid files into an image file which we can then
read parts from and copy onto the screen.

This is also a template for how we can implement other default typefaces
that may come with other platforms we may wish to emulate (for example,
C64).