Display @ symbols on startup
There's actually a lot going on here to make the @ symbol display
happen.

The @ symbols are the visual representation of NUL-character bytes. They
are inverse video, as dictated by the Apple //e technical reference. As
part of the bootup sequence, those bytes are overwritten with space
characters.

As part of the work, we now have a table of rows and columns that map to
display buffer addresses.